move the utilities to their own home
