+++ app/src/docs/components/Search/index.jsx
@@ -12,67 +12,12 @@ import useLunr from '$docs/hooks/useLunr'
 import useGetIndexStore from '$docs/hooks/useGetIndexStore'
 import RawHtml from '$shared/components/RawHtml'
 import BodyClass from '$shared/components/BodyClass'
+import { formatSearchResults } from './searchUtils'
 
 import styles from './search.pcss'
 
 type Props = {
     toggleOverlay: () => void,
-}
-
-const SNIPPET_CHARS = 210
-const MIN_MATCH_BUFFER_CHARS = 50
-const CHARS_AFTER_MATCH = SNIPPET_CHARS - MIN_MATCH_BUFFER_CHARS
-
-const generateTextSnippet = (text, matchPosition) => {
-    // A) No trimming
-    if (text.length < SNIPPET_CHARS) {
-        return text
-    }
-
-    // A) Trim end
-    if (matchPosition < MIN_MATCH_BUFFER_CHARS) {
-        return `${text.slice(0, SNIPPET_CHARS - 3)}...` //  - 3 for 3 ...
-    }
-
-    // A) Trim start & end
-    const startChar = matchPosition - (MIN_MATCH_BUFFER_CHARS - 3) //  - 3 for 3 ...
-    const endChar = matchPosition + (CHARS_AFTER_MATCH - 3) //  - 3 for 3 ...
-    return `...${text.slice(startChar, endChar)}...`
-}
-
-const highlightMatch = (text, startHighlightChar, endHighlightChar) => {
-    let textWithHighlight = text
-
-    textWithHighlight = [
-        text.slice(0, endHighlightChar),
-        '</span>',
-        text.slice(endHighlightChar),
-    ].join('')
-
-    textWithHighlight = [
-        textWithHighlight.slice(0, startHighlightChar),
-        '<span class="highlight">',
-        textWithHighlight.slice(startHighlightChar),
-    ].join('')
-
-    return textWithHighlight
-}
-
-const formatSearchResults = (results: Object) => {
-    if (!results.length) { return [] }
-
-    return results.slice(0, 10).map((result) => {
-        const { content: fullText, matchData } = result
-        const [matchPosition, matchLength] = matchData
-
-        const highlightedFullText = highlightMatch(fullText, matchPosition, matchPosition + matchLength).replace(/\n/g, ' ')
-        const textSnippet = generateTextSnippet(highlightedFullText, matchPosition)
-
-        return {
-            ...result,
-            textSnippet,
-        }
-    })
 }
 
 const Search = ({ toggleOverlay }: Props) => {
