+++ app/src/docs/hooks/useLunr.js
@@ -1,66 +1,7 @@
 // @flow
 
 import { useState, useEffect, useMemo } from 'react'
-import lunr from 'lunr'
-
-const formatMatchData = (matchData) => {
-    // matchPosition: The number of characters into the string where the match is located.
-    // matchLength: The number of characters of the keyword matched.
-    let matchPosition = 0
-    let matchLength = 0
-
-    Object.values(matchData).forEach((match: Object, matchIndex: number) => {
-        // highlighting first match only (for now)
-        if (matchIndex === 0) {
-            const matches = Object.values(match)
-            matches.forEach(({ content }: Object, contentIndex) => {
-                const { position } = content
-                if (contentIndex === 0 && content && position) {
-                    [[matchPosition, matchLength]] = position
-                }
-            })
-        }
-    })
-    return [matchPosition, matchLength]
-}
-
-function runQuery({ query, index, store }) {
-    if (!query || !index || !store) { return [] }
-
-    // 'trimmer' causes unexplainable crashes while searching.
-    lunr(function removeTrimmer() {
-        this.pipeline.remove(lunr.trimmer)
-    })
-
-    const lunrResults = () => (
-        // Search on the 'content' field of the array only.
-        // Store is built with the title as the first sentence of the 'content' field.
-        // Exact matches get a boost. Trailing wildcard (type-ahead) matches get a lower score.
-        index.query((q) => {
-            q.term(lunr.tokenizer(query), {
-                fields: ['content'],
-                boost: 100,
-                usePipeline: false,
-            })
-            q.term(lunr.tokenizer(query), {
-                fields: ['content'],
-                boost: 10,
-                usePipeline: false,
-                wildcard: lunr.Query.wildcard.TRAILING,
-            })
-        })
-    )
-
-    const results = lunrResults().map(({ ref, matchData }) => {
-        // The results are generated from the combination of combining the index with the store.
-        // matchData contains the position indexes of the matched keywords
-        const searchResults = store[ref]
-        searchResults.matchData = formatMatchData(matchData)
-        return searchResults
-    })
-
-    return results
-}
+import { runSearchQuery } from '../components/Search/searchUtils'
 
 const useLunr = (query: string, providedIndex: any, providedStore: any) => {
     const [index, setIndex] = useState(null)
@@ -72,7 +13,7 @@ const useLunr = (query: string, providedIndex: any, providedStore: any) => {
         setStore(providedStore)
     }, [providedIndex, providedStore])
 
-    return useMemo(() => runQuery({
+    return useMemo(() => runSearchQuery({
         query, index, store,
     }), [query, store, index])
 }
